refactor: Replace errors module with legacy_errors and update references throughout the codebase
