+++ src/market_data_ingestor/src/io/legacy_errors.rs
