+++ src/market_data_ingestor/src/lib.rs
@@ -1,5 +1,5 @@
 #[cfg(feature = "alpaca-python-sdk")]
-use errors::IngestorError;
+use legacy_errors::IngestorError;
 #[cfg(feature = "alpaca-python-sdk")]
 use requests::historical::StockBarData;
 #[cfg(feature = "alpaca-python-sdk")]
@@ -7,7 +7,7 @@ use utils::python_init::Config;
 
 #[cfg(feature = "cli")]
 pub mod cli;
-pub mod errors;
+pub mod legacy_errors;
 pub mod io;
 pub mod models;
 pub mod providers;
