+++ src/market_data_ingestor/src/requests/historical/legacy_errors.rs
